+++ blender-qmm/mats/base/Titanium.py
@@ -35,7 +35,7 @@ class QMMTitanium(bpy.types.Operator):
         # CreateShader
         m_titanium = bpy.data.materials.new(name="QMM Titanium Textured")
         m_titanium.use_nodes = True
-        m_titanium.diffuse_color = (0.337163, 0.296138, 0.254152, 1)
+        m_titanium.diffuse_color = (0.533276, 0.491021, 0.439657, 1)
         m_titanium.metallic = 1
         m_titanium.roughness = 0.35
 
@@ -49,7 +49,7 @@ class QMMTitanium(bpy.types.Operator):
         BSDF = nodes.get('Principled BSDF')
         BSDF.distribution = 'MULTI_GGX'
         BSDF.location = (-300, 0)
-        BSDF.inputs[0].default_value = (0.337163, 0.296138, 0.254152, 1)
+        BSDF.inputs[0].default_value = (0.533276, 0.491021, 0.439657, 1)
         BSDF.inputs[6].default_value = 1
         BSDF.inputs[9].default_value = 0.35
         # BSDF.inputs[16].default_value = 2.42
@@ -60,10 +60,10 @@ class QMMTitanium(bpy.types.Operator):
         ec_group.name = "Energy Conservation v5"
         ec_group.node_tree = bpy.data.node_groups['Energy Conservation v5']
         ec_group.location = (-500, -200)
-        ec_group.inputs[0].default_value = (0.337163, 0.296138, 0.254152, 1)
+        ec_group.inputs[0].default_value = (0.533276, 0.491021, 0.439657, 1)
         ec_group.inputs[1].default_value = 0.35
         # ec_group.inputs[2].default_value = 2.42
-        ec_group.inputs[4].default_value = (0.428690, 0.423267, 0.428690, 1)
+        ec_group.inputs[4].default_value = (0.686685, 0.679542, 0.686685, 1)
 
         # TexturizerGroup
         bpy.ops.node.texturizer_group_operator()