+++ blender-qmm/mats/noble/Silver.py
@@ -60,10 +60,10 @@ class QMMSilver(bpy.types.Operator):
         # EnergyConservationGroup
         bpy.ops.node.ec_group_operator()
         ec_group = self.make_node(nodes, "Energy Conservation v5", 'Energy Conservation v5', -500, -200)
-        ec_group.inputs[0].default_value = (0.913098, 0.879622, 0.830770, 1)
+        ec_group.inputs[0].default_value = (0.964685, 0.947306, 0.921582, 1)
         ec_group.inputs[1].default_value = 0.075
         # ec_group.inputs[2].default_value = 1.57
-        ec_group.inputs[4].default_value = (0.991101, 0.991101, 0.991102, 1)
+        ec_group.inputs[4].default_value = (1, 1, 1, 1)
         links(ec_group.outputs[0], BSDF.inputs[0])
         links(ec_group.outputs[1], BSDF.inputs[7])
         links(ec_group.outputs[2], BSDF.inputs[9])