+++ blender-qmm/mats/noble/Gold.py
@@ -60,7 +60,8 @@ class QMMGold(bpy.types.Operator):
         ec_group.inputs[0].default_value = (0.947306, 0.775822, 0.371238, 1)
         ec_group.inputs[1].default_value = 0.075
         # ec_group.inputs[2].default_value = 1.45
-        ec_group.inputs[4].default_value = (1.000000, 0.982250, 0.752942, 1)
+        # ec_group.inputs[4].default_value = (1.000000, 0.982250, 0.752942, 1)
+        ec_group.inputs[4].default_value = [1.040075, 1.111826, 1.486759, 1]
         ec_group.location = (-500, -200)
         links(ec_group.outputs[0], BSDF.inputs[0])
         links(ec_group.outputs[1], BSDF.inputs[7])